+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - List can be a list of GranuleURs or granule concept-IDs
 - Update db size from t2.micro to t3.micro
 - Made arguments *--cumulus-configurations* and *--default-message-config* optional in preview mode
+- Added check before backfilling images--make sure it is enabled in forge-tig configuration
 - Updated forge-py to 0.4.0
 - Updated cumulus-postworkflow-normalizer to 0.4.1
 - Updated hitide-backfill-lambdas to 0.4.1